requirements.txt
streamlit
agno==1.1.1
python-dotenv
markdown

# Gemini (Google LLM)
google-genai==1.41.0

# Web search tools used by Agno
google-search-results==2.4.2
googlesearch-python==1.3.0
